Remplace le script par un test jest
